pyproject.toml
[tool.poetry]
name = "dj_haystack"
version = "0.1.0"
description = ""
authors = ["surajmt8848 <suraj8848mt@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.3"
djangorestframework = "^3.12.4"
isort = "^5.8.0"
django-haystack = "^3.0"
django-autoslug = "^1.9.8"
drf-haystack = "^1.8.10"
django-haystack-elasticsearch = "^0.1.0"
elasticsearch = "^5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
